Se renombra el proyecto a ms-persistence
