+++ ms-persistence/pom.xml
@@ -12,7 +12,7 @@
     </parent>
 
     <groupId>com.nttdata.bootcamp</groupId>
-    <artifactId>ms-database</artifactId>
+    <artifactId>ms-persistence</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
