+++ ms-persistence/src/main/java/com/nttdata/bootcamp/mspersistence/MsDatabaseApplication.java
@@ -1,4 +1,4 @@
-package com.nttdata.bootcamp.msdatabase;
+package com.nttdata.bootcamp.mspersistence;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
