+++ ms-persistence/src/main/java/com/nttdata/bootcamp/mspersistence/application/ClientService.java
@@ -1,4 +1,4 @@
-package com.nttdata.bootcamp.ms.client.application;
+package com.nttdata.bootcamp.mspersistence.application;
 
 import com.nttdata.bootcamp.ms.commons.base.domain.ClientDTO;
 import reactor.core.publisher.Flux;
@@ -7,9 +7,13 @@ import reactor.core.publisher.Mono;
 public interface ClientService {
 
     Mono<ClientDTO> create(ClientDTO clientDTO);
+
     Mono<ClientDTO> findById(String code);
+
     Flux<ClientDTO> listAll();
+
     void update(ClientDTO clientDTO);
+
     void delete(String code);
 
 }
