+++ ms-persistence/src/main/java/com/nttdata/bootcamp/mspersistence/infraestructure/AccountRepository.java
@@ -1,6 +1,6 @@
-package com.nttdata.bootcamp.msdatabase.infraestructure;
+package com.nttdata.bootcamp.mspersistence.infraestructure;
 
-import com.nttdata.bootcamp.msdatabase.model.Account; 
+import com.nttdata.bootcamp.mspersistence.model.Account;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
 public interface AccountRepository extends ReactiveMongoRepository<Account, String> {
