+++ ms-persistence/src/main/java/com/nttdata/bootcamp/mspersistence/infraestructure/ClientRepository.java
@@ -1,6 +1,6 @@
-package com.nttdata.bootcamp.msdatabase.infraestructure;
+package com.nttdata.bootcamp.mspersistence.infraestructure;
 
-import com.nttdata.bootcamp.msdatabase.model.Client;
+import com.nttdata.bootcamp.mspersistence.model.Client;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
 public interface ClientRepository extends ReactiveMongoRepository<Client, String> {
