--- ms-transaction/src/main/java/com/nttdata/bootcamp/Main.java
@@ -1,7 +1,0 @@
-package com.nttdata.bootcamp;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
