+++ ms-transaction/src/main/java/com/nttdata/bootcamp/ms/transaction/MsTransactionApplication.java
@@ -1,0 +1,12 @@
+package com.nttdata.bootcamp.ms.transaction;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MsTransactionApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(MsTransactionApplication.class, args);
+    }
+
+}
