+++ ms-transaction/src/main/java/com/nttdata/bootcamp/ms/transaction/application/TransactionService.java
@@ -1,19 +1,15 @@
-package com.nttdata.bootcamp.msdatabase.application;
+package com.nttdata.bootcamp.ms.transaction.application;
 
 import com.nttdata.bootcamp.ms.commons.base.domain.ClientDTO;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public interface ClientService {
+public interface TransactionService {
 
     Mono<ClientDTO> create(ClientDTO clientDTO);
-
     Mono<ClientDTO> findById(String code);
-
     Flux<ClientDTO> listAll();
-
     void update(ClientDTO clientDTO);
-
     void delete(String code);
 
 }
